Layouting: The first stuff: Functions now have depencies to each other, so it is safe now to change element dimensions from within a function. PrepareRuns are executed until all depencies are resolved.
BugFix: PDFTarget's FontFactory did not prevent the BaseFont of iText from caching.
This can lead to OutOfMemoryErrors if a large amount of TrueTypeFonts are installed.

Reorganisation of ReportStates, now an onw package, more to do here...
Expressions and Functions are contained within the same structure while processing the
report (LeveledExpressionList)...