Fixed a couple of conversion issues:
All conversions from an to a number or date value will never return null.
If they fail, they fail with a TypeConversionException. I removed a lot
of "value == null" and NumberFormatException-Catches from the code now.

On numeric values, the Exact-Function performed an int-comparation, which
is invalid. It must perform a rounded double comparision instead.

The DefaultComparator was invalid.
The conversion from text to dates was invalid.